% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plan.R
\name{analysis_plan}
\alias{analysis_plan}
\title{Create analysis plan}
\usage{
analysis_plan(countries)
}
\arguments{
\item{countries}{character vector of countries}
}
\value{
drake plan object
}
\description{
Function to create a drake plan for a basic analysis
}
\examples{

\dontrun{
library(gapminder)
library(dplyr)
library(drake)

my_plan <- gapminder \%>\%
  distinct(country) \%>\%
  unlist() \%>\% unname() \%>\%
  as.character() \%>\%
  analysis_plan()

make(my_plan)

readd(results)
readd(plot)
}
}
