% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plan_api.R
\name{analysis_plan_api}
\alias{analysis_plan_api}
\title{Create API analysis plan}
\usage{
analysis_plan_api(countries, url = "http://localhost:8000/",
  path = "model")
}
\arguments{
\item{countries}{character vector of countries}
}
\value{
drake plan object
}
\description{
Function to create a drake plan for an analysis using the API
}
\examples{

\dontrun{
library(gapminder)
library(dplyr)
library(drake)

# note will have to have instantiated the API in another R process eg:
# gapminder_api(8000)

my_plan <- gapminder \%>\%
  distinct(country) \%>\%
  unlist() \%>\% unname() \%>\%
  as.character() \%>\%
  analysis_plan_api()

make(my_plan)

readd(results)
readd(plot)
}
}
