% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_coef.R
\name{calculate_coef}
\alias{calculate_coef}
\title{Calculate coeficients}
\usage{
calculate_coef(country = NULL, df = gapminder::gapminder, sleep = 0)
}
\arguments{
\item{country}{country to estimate}

\item{df}{data frame in gapminder format}

\item{sleep}{time to sleep}
}
\value{
data frame
}
\description{
Calculate coeficients
}
\examples{
calculate_coef('New Zealand')
}
