% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_coef_api.R
\name{calculate_coef_api}
\alias{calculate_coef_api}
\title{Calculate coefficients from the API}
\usage{
calculate_coef_api(country = NULL, url = "http://localhost:8001/",
  path = "model")
}
\arguments{
\item{country}{country to fit the model to}

\item{url}{URL of API server}

\item{path}{endpoint of API}
}
\value{
data frame
}
\description{
Calculate coefficients from the API
}
\examples{
\dontrun{
calculate_coef_api('Algeria', url='http://localhost:8001/', path='model')
}
}
