% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapminder_api.R
\name{gapminder_api}
\alias{gapminder_api}
\title{Gapminder Pipeline API}
\usage{
gapminder_api(port = 8001)
}
\arguments{
\item{port}{Define port to serve API on - default is 8001}
}
\value{
deploys a plumber API on port 8001
}
\description{
Gapminder Pipeline API
}
\examples{
NULL
}
