% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coef.R
\name{plot_coef}
\alias{plot_coef}
\title{Plot coefficients}
\usage{
plot_coef(df)
}
\arguments{
\item{df}{data frame generated by broom::tidy.lm}
}
\value{
ggplot2 object
}
\description{
Plot coefficients
}
\examples{
NULL
}
