% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_country.R
\name{plot_country}
\alias{plot_country}
\title{Plot data for one or more countries}
\usage{
plot_country(country = NULL, df = gapminder::gapminder)
}
\arguments{
\item{country}{character vector of countries}

\item{df}{gapminder data frame}
}
\value{
ggplot2 object
}
\description{
Plot data for one or more countries
}
\examples{
NULL
}
