% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slow_lm.R
\name{slow_lm}
\alias{slow_lm}
\title{Slow Linear Model fit}
\usage{
slow_lm(df, sleep = 0)
}
\arguments{
\item{df}{data in gapminder format}

\item{sleep}{}
}
\value{
lm fit object
}
\description{
Slow Linear Model fit
}
\examples{
gapminder::gapminder \%>\%
 dplyr::filter(country == 'New Zealand') \%>\%
 slow_lm()
}
