% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_test_viz.R
\name{test_viz}
\alias{test_viz}
\title{Testviz shiny app}
\usage{
test_viz()
}
\value{
runs a shiny app
}
\description{
Convenience function to run the test_viz shiny app within an R session.
}
\details{
The test_viz shiny app allows the user to view gapminder data for a particular country.
}
\examples{
\dontrun{

test_viz()

}
}
